render_context: output lost signal for surface to reassign outputs on output lost
