+++ include/taiwins/render_context.h
@@ -103,6 +103,7 @@ struct tw_render_context {
 		struct wl_signal dma_set;
 		struct wl_signal compositor_set;
 		//this is plain damn weird.
+		struct wl_signal output_lost;
 		struct wl_signal wl_surface_dirty;
 		struct wl_signal wl_surface_destroy;
 	} events;
@@ -124,6 +125,9 @@ tw_render_context_destroy(struct tw_render_context *ctx);
 void
 tw_render_context_build_view_list(struct tw_render_context *ctx,
                                   struct tw_layers_manager *manager);
+void
+tw_render_surface_reassign_outputs(struct tw_render_surface *render_surface,
+                                   struct tw_render_context *ctx);
 void
 tw_render_context_set_dma(struct tw_render_context *ctx,
                           struct tw_linux_dmabuf *dma);
